double

+++ PhysicalNumber.cpp
@@ -5,7 +5,7 @@ using namespace std;
 using namespace ariel;
 //2 km + 300 m 300*100 = 30000/100000 = 2.3
 //300m + 2km =2*100000 =200000/100 = 2300
-const long value[9]={1, 100, 100000, 1, 60, 3600, 1, 1000, 1000000};
+const double value[9]={1, 100, 100000, 1, 60, 3600, 1, 1000, 1000000};
 const string name[9]={"cm", "m", "km", "sec", "min", "hour", "g", "kg", "ton"};
 PhysicalNumber::PhysicalNumber(double x1,Unit U1){
         x=x1;
@@ -67,14 +67,14 @@ PhysicalNumber& PhysicalNumber::operator=(const PhysicalNumber& n){
 const PhysicalNumber PhysicalNumber::operator+(const PhysicalNumber& n) const {
         if(someType(n)==false) {  __throw_invalid_argument("Not some type");}
         else{
-                long double answer = this->x + (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
+                 double answer = this->x + (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
                 return PhysicalNumber(answer,this->U);
         }
 }
 const PhysicalNumber PhysicalNumber::operator-(const PhysicalNumber& n) const{
         if(someType(n)==false) {  __throw_invalid_argument("Not some type");}
         else{
-                long double answer = this->x - (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
+                 double answer = this->x - (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
                 return PhysicalNumber(answer,this->U);
         }
 }
@@ -118,7 +118,7 @@ istream& ariel::operator>>(istream &is, PhysicalNumber& n){
 const bool PhysicalNumber::operator==(const PhysicalNumber& n) const{
         if(someType(n)==false) {  __throw_invalid_argument("Not some type");}
         else{
-                long double answer =  (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
+                 double answer =  (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
                 return x==answer;
         }
 
@@ -126,7 +126,7 @@ const bool PhysicalNumber::operator==(const PhysicalNumber& n) const{
 const bool PhysicalNumber::operator<(const PhysicalNumber& n) const{
         if(someType(n)==false) {  __throw_invalid_argument("Not some type");}
         else{
-                long double answer =  (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
+                 double answer =  (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
                 return x<answer;
         }
 
@@ -134,28 +134,28 @@ const bool PhysicalNumber::operator<(const PhysicalNumber& n) const{
 const bool PhysicalNumber::operator>(const PhysicalNumber& n) const{
         if(someType(n)==false) {  __throw_invalid_argument("Not some type");}
         else{
-                long double answer = (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
+                 double answer = (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
                 return x>answer;
         }
 }
 const bool PhysicalNumber::operator<=(const PhysicalNumber& n) const{
         if(someType(n)==false) {  __throw_invalid_argument("Not some type");}
         else{
-                long double answer = (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
+                 double answer = (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
                 return x<=answer;
         }
 }
 const bool PhysicalNumber::operator>=(const PhysicalNumber& n) const{
         if(someType(n)==false) {  __throw_invalid_argument("Not some type");}
         else{
-                long double answer =  (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
+                 double answer =  (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
                 return x>=answer;
         }
 }
 const bool PhysicalNumber::operator!=(const PhysicalNumber& n) const{
         if(someType(n)==false) {  __throw_invalid_argument("Not some type");}
         else{
-                long double answer = (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
+                 double answer = (n.x * (double(value[(int)n.U]) / value[(int)this->U]));
                 return x!=answer;
         }
 }
